refactor(gateway): name connector target module explicitly
